move SVM validation dataset to be implicitly defined in the trainer's datamodule
this fixes a bug where the SVM dataset was actually just a copy of the pytorch lightning trainer's dataset being used for overall model training/validation. this means that I could not change dataset hyperparameters between the svm and regular datasets (i.e. returning `semantic_id` in one but not the other). this does not effect previous results.