allow for custom group scaling radius different from ball query radius.
this is useful for when you want to try to have a pretrained work with smaller groups. since the original method scales the groups to a unit radius no matter what, the scale of the groups would change after changing these sizes. to keep them the same, the user should set `rescale_by_group_radius` to the original radius that the model was pre-trained on, and `group_radius` to the new one.